Corrects credential handling, takes into consideration user state changes
