+++ src/index.js
@@ -10,18 +10,6 @@ import configureStore from './state/store/configureStore'
 import { verifyCredentials } from './state/actions/reduxTokenAuthConfig'
 import axios from 'axios'
 
-const getCurrentCredentials = () => {
-  const credentials = {
-    "access-token": localStorage.getItem("access-token"),
-    "token-type": localStorage.getItem("token-type"),
-    client: localStorage.getItem("client"),
-    expiry: localStorage.getItem("expiry"),
-    uid: localStorage.getItem("uid")
-  }
-  return credentials
-}
-
-axios.defaults.headers = getCurrentCredentials()
 
 const store = configureStore()
